requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
APScheduler==3.10.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bilibili-api-python==15.5.1
blinker==1.5
Brotli==1.0.9
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.6
colorama==0.4.6
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.2
httpx==0.24.1
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.3
multidict==6.0.4
Pillow==9.5.0
priority==2.0.0
pyasn1==0.5.0
pycryptodomex==3.18.0
pyee==11.0.0
pypng==0.20220715.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
quart==0.18.4
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
tqdm==4.65.0
typing_extensions==4.7.0
tzlocal==5.0.1
urllib3==2.0.3
Werkzeug==2.3.6
wsproto==1.2.0
yarl==1.9.2
